obfuscate passwords and credentials in the gui (literal "shadow passwords!")
applies text shadowing to blur keys that say obvious things like "password" and "credential",
to config table and sensors table. clears up when you click it.
this prevents people looking over your shoulder from seeing things they shouldn't,
but it doesn't block REST access, and if you click on it you can still see it.
(this is a common trick done at AWS & SL, btw.)

a separate feature is to enforce visibility of sensors; this can be done with entitlements on a per-sensor basis
but it might be nice to have an easy entitlements mode where "sensitive" info is not available,
and options on config keys (similar to how i just did it with ConfigInheritance,
in https://github.com/apache/incubator-brooklyn/pull/483) to allow ConfigSensitivity.