Revert "Delegate pax-web-spi version to dependencyManagement in parent pom"
This reverts commit c5692a2ec5c0bb0409e0369776556e3794b09f5a.

When building the base brooklyn project, i.e. that which contains
brooklyn-server and brooklyn-ui as submodules, then the maven project fails
validation. This is because the brooklyn-ui project doesn't include a
relativePath element pointing to the pom inside the brookly-server project.

A note in the brooklyn-ui pom explains why there should not be a relativePath
element to it's sibling (../brooklyn-server). Thus the maven reactor for
brooklyn-ui uses whatever brooklyn-parent is available in the reactor path (i.e.
~/.m2/repository, or maven central, etc). I will not use the 'current'
development copy in ../brooklyn-server/parent.