remove 'verify' phase
verify is failing with warnings about "warn" not being found (using new karaf-plugin version);
don't think we changed anything to cause that and the resulting assembly is fine.
not sure verify did much anyway!

this commit also bring versions and config for that plugin in synch with what is done elsewhere.