Revert "add jetty-web using sharing session manager"
This reverts commit 7c4eb4c797e2946fd47a7c2fce2d9a156e549d12.

The filter installs a shared session manager so no need for us to do it here.
Kept as a commit in the history however as it is cleaner; a better way would be
to define a session handler in the REST bundle and install it here, so all the
webapps have a shared handler that outlives them.