improve how yaml generated for "open in composer"
previously if there was a name set on the app and one set in the ui (and if there was a name set on the app the ui always inferred one!),
or any other duplicated keys etc, the yaml generated would duplicate the key and thus fail to load.

now the code tries to dedupe those fields.  it prefers the supplied yaml, in case there are comments etc,
and is smart about where/what it updates.

if errors are detected, the original yaml is preserved and any new fields prepended, with a warning,
and yaml text editor mode is targeted in the composer.  this helps with the case where the template deliberately contains comments or errors to be addressed.