use ** pattern in rat exclude rules
seems to be needed in some cases if folder name is different; there might be a better portable way to do current directory but the things we exclude can be everywhere.
not sure why the children patterns aren't being picked up but declaring it at root also seems to work. (the problem happens when we link to brooklyn-ui in a nested/skinning project.)